Add photos controller to handle all /photo related requests.
