+++ server/index.js
@@ -11,27 +11,10 @@ const passport = require('passport');
 const _500pxStrategy = require('passport-500px').Strategy;
 const cookieParser = require('cookie-parser');
 const session = require('express-session');
-const oauthSignature = require('oauth-signature');
+const CONSTANTS = require('./constants');
 
-// helpers
-const generateRandomString = require('./helpers/generateRandomString');
-const constructOAuthUrl = require('./helpers/constructOAuthUrl');
-const constructAuthorizationHeader = require('./helpers/constructAuthorizationHeader');
-
-const CONSUMER_KEY = process.env.CONSUMER_KEY;
-const CONSUMER_SECRET = process.env.CONSUMER_SECRET;
-
-// URLs
-const ROOT_URL = 'https://api.500px.com/v1/photos';
-const LIKE_PHOTO_URL_SETTINGS = '/vote?vote=1';
-const CONSUMER_KEY_SETTING = `?consumer_key=${CONSUMER_KEY}&`;
-const GET_PHOTOS_URL_SETTINGS = 'feature=popular&sort=created_at&image_size=3&include_store=store_download&include_states=voted&rpp=100';
-const CALLBACK_URL = 'http://localhost:3000/login/500px/callback/';
-const CLIENT_URL = 'http://localhost:8080';
-
-
-// Constructed URL
-const API_URL_GET_PHOTOS = `${ROOT_URL}${CONSUMER_KEY_SETTING}${GET_PHOTOS_URL_SETTINGS}`;
+// Controllers
+const PhotosController = require('./controllers/PhotosController');
 
 /*
   Passport setup
@@ -46,9 +29,9 @@ passport.deserializeUser(function(obj, done) {
 });
 
 passport.use(new _500pxStrategy({
-    consumerKey: CONSUMER_KEY,
-    consumerSecret: CONSUMER_SECRET,
-    callbackURL: CALLBACK_URL,
+    consumerKey: CONSTANTS.CONSUMER_KEY,
+    consumerSecret: CONSTANTS.CONSUMER_SECRET,
+    callbackURL: CONSTANTS.CALLBACK_URL,
   },
   function(token, tokenSecret, profile, done) {
     done(null, profile);
@@ -60,7 +43,6 @@ passport.use(new _500pxStrategy({
 */
 
 const app = express();
-
 
 app.use(morgan('combined'));
 app.use(cors());
@@ -76,80 +58,16 @@ app.use(session({
 }));
 app.use(passport.initialize());
 app.use(passport.session());
-app.use(express.static(__dirname + '/public'));
 
 // Routes
 app.get('/', function(req, res) {
   res.send({ message: 'success!' });
 });
 
-app.get('/photos', function(req, res) {
-  request(API_URL_GET_PHOTOS, function(error, response, body) {
-    if (error) {
-      return res.send({ error: error });
-    }
 
-    if (!response) {
-      return res.send({ error: 'error', statusCode: 'Unknown' });
-    }
-
-    if (response.statusCode == 200) {
-      return res.send(body);
-    }
-
-    // only reaches here if the statusCode is not 200 - for error handling
-    res.send(body);
-  });
-});
-
-app.post('/photos/:id/vote', function(req, res) {
-  const photoId = req.params.id;
-  const oauthToken = res.req.body.oauth_token;
-  const baseUrl = `${ROOT_URL}/${photoId}${LIKE_PHOTO_URL_SETTINGS}`
-
-  const params = {
-    oauth_token: oauthToken,
-    oauth_consumer_key: CONSUMER_KEY,
-    oauth_signature_method: 'HMAC-SHA1',
-    oauth_timestamp: Date.now(),
-    oauth_nonce: generateRandomString(),
-    oauth_version: '1.0',
-  };
-
-  // generates oauth_signature
-  // null field is optional token_secret
-  const signature = oauthSignature.generate('POST', baseUrl, params, CONSUMER_SECRET);
-
-  params['oauth_signature'] = signature;
-
-  // const constructedUrl = constructOAuthUrl(baseUrl, params);
-  const authorizationHeader = constructAuthorizationHeader(params);
-
-  const options = {
-    url: baseUrl,
-    headers: {
-      Authorization: authorizationHeader,
-    },
-  };
-
-  // Tried passing in either the options or constructedUrl
-  // as the first argument below.
-  // Passing in options (with the signature containing 'null' for the token_secret above),
-  // cause a 500 error.
-  // Passing in the constructedUrl causes a 401 error.
-  request.post(options, function(err, response, body) {
-    if (err) {
-      console.log('ERROR!!!!', err);
-      res.send({ error: err });
-    }
-
-    if (response.statusCode === 200) {
-      res.send(body);
-    }
-
-    res.send(response);
-  });
-});
+// Photos
+app.get('/photos', PhotosController.getPhotos);
+app.post('/photos/:id/vote', PhotosController.likePhoto);
 
 
 // OAuth route
@@ -167,7 +85,7 @@ app.get('/login/500px/callback',
   function(req, res) {
     // console.log('REQUEST', req);
 
-    res.redirect(`${CLIENT_URL}?oauth_token=${req.query.oauth_token}`);
+    res.redirect(`${CONSTANTS.CLIENT_URL}?oauth_token=${req.query.oauth_token}`);
   }
 );
