+++ server/package.json
@@ -16,6 +16,7 @@
     "dotenv": "^4.0.0",
     "express": "^4.14.1",
     "express-session": "^1.15.0",
+    "mongoose": "^4.8.2",
     "morgan": "^1.7.0",
     "nodemon": "^1.11.0",
     "oauth-signature": "^1.3.1",
